+++ scrapy_poet/__init__.py
@@ -1,4 +1,4 @@
-from .api import DummyResponse, callback_for
+from .api import AnnotatedResult, DummyResponse, callback_for
 from .downloadermiddlewares import InjectionMiddleware
 from .page_input_providers import HttpResponseProvider, PageObjectInputProvider
 from .spidermiddlewares import RetryMiddleware